Renaming and refactoring of cache, parallel and trajectory
cache -> mesh_cache
parallel -> parallel_render
trajectory -> camera_trajectory
