+++ znvis/camera_trajectories/base_trajectory.py
