+++ znvis/parallel_render/parallel_render_worker.py
@@ -6,7 +6,7 @@ import pickle
 import sys
 import traceback
 
-from znvis.visualizer.parallel.parallel_render_manager import (
+from znvis.parallel_render.parallel_render_manager import (
     _initialize_parallel_worker,
     _render_frame_parallel_worker,
 )
